+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
   - Add Arguments.setPreparedArgumentsEnabled to disable new preparable arguments feature
   - Add JUnit 5 support to the testing package.
   - Add interceptors for type inference for Row Mappers, Column Mappers and SQL Array types
+  - Fix mapper lookup for Kotlin data types. Ensure that registered prefixes are used. Fixes #1944
 
 # 3.23.0
   - Rework and document importing bindings in guice definition modules